// src/common/enums/HttpMessage.enum.ts
// src/http-message.enum.ts
export enum HttpMessage {
    //Custom
    EMAIL_NOT_VERIFY = 'Correo eléctronico no verificado.',
    INVALID_USER_EMAIL = 'Usuario o correo eléctronico inválido.',
    INVALID_PASSWORD = 'Contraseña inválida.',
    INVALID_LOGIN = 'Credenciales inválidas.',

    // Informational responses (100–199)
    CONTINUE = 'Continuar',
    SWITCHING_PROTOCOLS = 'Cambiando protocolos',
    PROCESSING = 'Procesando',
    EARLY_HINTS = 'Primeros indicios',

    // Successful responses (200–299)
    OK = 'Operación exitosa',
    CREATED = 'Registro creado correctamente',
    ACCEPTED = 'Aceptado',
    NON_AUTHORITATIVE_INFORMATION = 'Información no autoritativa',
    NO_CONTENT = 'Sin contenido',
    RESET_CONTENT = 'Restablecer contenido',
    PARTIAL_CONTENT = 'Contenido parcial',
    MULTI_STATUS = 'Multi-estado',
    ALREADY_REPORTED = 'Ya reportado',
    IM_USED = 'Estoy usado',

    // Redirection messages (300–399)
    MULTIPLE_CHOICES = 'Múltiples opciones',
    MOVED_PERMANENTLY = 'Movido permanentemente',
    FOUND = 'Encontrado',
    SEE_OTHER = 'Ver otro',
    NOT_MODIFIED = 'No modificado',
    USE_PROXY = 'Usar proxy',
    TEMPORARY_REDIRECT = 'Redirección temporal',
    PERMANENT_REDIRECT = 'Redirección permanente',

    // Client error responses (400–499)
    BAD_REQUEST = 'Solicitud incorrecta',
    UNAUTHORIZED = 'No autorizado',
    PAYMENT_REQUIRED = 'Pago requerido',
    FORBIDDEN = 'Prohibido',
    NOT_FOUND = 'No encontrado',
    METHOD_NOT_ALLOWED = 'Método no permitido',
    NOT_ACCEPTABLE = 'No aceptable',
    PROXY_AUTHENTICATION_REQUIRED = 'Autenticación de proxy requerida',
    REQUEST_TIMEOUT = 'Tiempo de espera agotado',
    CONFLICT = 'Conflicto',
    GONE = 'Desaparecido',
    LENGTH_REQUIRED = 'Longitud requerida',
    PRECONDITION_FAILED = 'Precondición fallida',
    PAYLOAD_TOO_LARGE = 'Carga útil demasiado grande',
    URI_TOO_LONG = 'URI demasiado larga',
    UNSUPPORTED_MEDIA_TYPE = 'Tipo de medio no soportado',
    RANGE_NOT_SATISFIABLE = 'Rango no satisfactorio',
    EXPECTATION_FAILED = 'Expectativa fallida',
    I_AM_A_TEAPOT = 'Soy una tetera',
    MISDIRECTED_REQUEST = 'Solicitud mal dirigida',
    UNPROCESSABLE_ENTITY = 'Entidad no procesable',
    LOCKED = 'Bloqueado',
    FAILED_DEPENDENCY = 'Dependencia fallida',
    TOO_EARLY = 'Demasiado temprano',
    UPGRADE_REQUIRED = 'Actualización requerida',
    PRECONDITION_REQUIRED = 'Precondición requerida',
    TOO_MANY_REQUESTS = 'Demasiadas solicitudes',
    REQUEST_HEADER_FIELDS_TOO_LARGE = 'Campos de encabezado de solicitud demasiado grandes',
    UNAVAILABLE_FOR_LEGAL_REASONS = 'No disponible por razones legales',

    // Server error responses (500–599)
    INTERNAL_SERVER_ERROR = 'Error interno del servidor',
    NOT_IMPLEMENTED = 'No implementado',
    BAD_GATEWAY = 'Mala pasarela',
    SERVICE_UNAVAILABLE = 'Servicio no disponible',
    GATEWAY_TIMEOUT = 'Tiempo de espera de la pasarela agotado',
    HTTP_VERSION_NOT_SUPPORTED = 'Versión HTTP no soportada',
    VARIANT_ALSO_NEGOTIATES = 'La variante también negocia',
    INSUFFICIENT_STORAGE = 'Almacenamiento insuficiente',
    LOOP_DETECTED = 'Bucle detectado',
    NOT_EXTENDED = 'No extendido',
    NETWORK_AUTHENTICATION_REQUIRED = 'Autenticación de red requerida',
}